test(core): Add functional test suite for slash command time tracking
ref: #1109 closes #1092
